+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Set button font-weight to 500, 600 for white text
 * Moved mui-colors.css to extra/ directory in dist/
 * Added mui-combined.js to extra/
+* Added mui-react-combined.js to extra/
 
 ## 0.1.15 - June 28, 2015